requirements.txt
# Core Data Science Libraries
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Data Fetching
yfinance==0.2.28
nsepy==0.8
requests==2.31.0
beautifulsoup4==4.12.2

# Technical Analysis
ta-lib==0.4.28
pandas-ta==0.3.14b0

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6
tensorflow==2.13.0
keras==2.13.1
prophet==1.1.4
statsmodels==0.14.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
mplfinance==0.12.9b7

# Backtesting
backtrader==1.9.78.123
vectorbt==0.25.3

# Options Pricing
py_vollib==1.0.1
mibian==0.1.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.65.0
joblib==1.3.1

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Jupyter & Notebooks
jupyter==1.0.0
notebook==7.0.2
ipywidgets==8.1.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Logging
loguru==0.7.0

# Database (optional)
sqlalchemy==2.0.19
pymongo==4.4.1

